+++ python/test/robot/test_joint_state.py
@@ -1,4 +1,6 @@
 import unittest
+import copy
+
 from state_representation import JointState
 
 JOINT_STATE_METHOD_EXPECTS = [
@@ -48,10 +50,20 @@ JOINT_STATE_METHOD_EXPECTS = [
 
 
 class TestJointState(unittest.TestCase):
+    def assert_np_array_equal(self, a, b):
+        self.assertListEqual(list(a), list(b))
+
     def test_callable_methods(self):
         methods = [m for m in dir(JointState) if callable(getattr(JointState, m))]
         for expected in JOINT_STATE_METHOD_EXPECTS:
             self.assertIn(expected, methods)
+
+    def test_copy(self):
+        state = JointState().Random("test", 3)
+        for state_copy in [copy.copy(state), copy.deepcopy(state)]:
+            self.assertEqual(state.get_name(), state_copy.get_name())
+            self.assertListEqual(state.get_names(), state_copy.get_names())
+            self.assert_np_array_equal(state.data(), state_copy.data())
 
 if __name__ == '__main__':
     unittest.main()