+++ python/test/space/test_spatial_state.py
@@ -1,7 +1,7 @@
 import unittest
+import copy
 
 from state_representation import SpatialState, StateType
-
 from test.test_state import STATE_METHOD_EXPECTS
 
 SPATIAL_STATE_METHOD_EXPECTS = [
@@ -38,6 +38,14 @@ class TestState(unittest.TestCase):
         self.assertEqual(state2.get_name(), state3.get_name())
         self.assertEqual(state2.get_reference_frame(), state3.get_reference_frame())
         self.assertEqual(state2.is_empty(), state3.is_empty())
+
+    def test_copy(self):
+        state = SpatialState(StateType.CARTESIANSTATE, "test", "robot", False)
+        for state_copy in [copy.copy(state), copy.deepcopy(state)]:
+            self.assertEqual(state.get_type(), state_copy.get_type())
+            self.assertEqual(state.get_name(), state_copy.get_name())
+            self.assertEqual(state.get_reference_frame(), state_copy.get_reference_frame())
+            self.assertEqual(state.is_empty(), state_copy.is_empty())
 
     def test_compatibility(self):
         state1 = SpatialState(StateType.CARTESIANSTATE, "test", "robot", True)